Area Under the Precision-Recall Curve for an Embedding

Embedding quality measure.

Usage

pr_auc(dm, labels)

Arguments

dm

Distance matrix of an embedding.

labels

Vector of labels for each observation in the dataset in the same order as the observations in the distance matrix.

Value

Area Under the Precision-Recall curve, averaged over each observation.

Details

The PR curve plots precision (also known as positive predictive value, PPV) against recall (also known as the true positive rate). The area under the curve provides similar information compared to the area under the ROC curve, but may be more appropriate when classes are highly imbalanced.

This function calculates the PR curve N times, where N is the number of the observations. The label of the Nth observation is set as the positive class and then the other observations are ranked according to their distance from the Nth observation in the output coordinates (lower distances being better). Observations with the same label as the Nth observation count as positive observations. The final reported result is the average over all observations. Rows with undefined AUC values, such as rows that cannot define both positive and negative examples, are excluded from overall and per-label averages. If no rows remain for an average, that average is NA_real_.

Perfect retrieval results in an AUC of 1. For random retrieval, the value is the proportion of the positive class labels for that curve.

Note

Use of this function requires that the PRROC package be installed.

References

Keilwagen, J., Grosse, I., & Grau, J. (2014). Area under precision-recall curves for weighted and unweighted data. PloS One, 9(3), e92209.

Davis, J., & Goadrich, M. (2006, June). The relationship between Precision-Recall and ROC curves. In Proceedings of the 23rd international conference on Machine learning (pp. 233-240). ACM.